requirements.txt
Django==2.0b1
IMDbPY==5.1.1
